README-hacking.md: minor clarifications in checklist
Roberto Baena recently tried building a new project with Maneage and
provided the following suggestions to make it more clear for a new user:

1) In the part where we talk about creating a Git repository, we should
highlight that it must be empty. This is because some (for example Gitlab)
propose to include a `README' file. But if the project is not empty, Git
will not allow pushing to it.

2) The `(can be done later)' comment was removed from the "Delete dummy
parts") to avoid confusion about applying some of them, but not others: if
only some are done, it may cause problems in the build.