README-hacking.md: described why automatic preparation only occurs once
Recently (since Commit 7d0c5ef77), the preparation is not run automatically
every time. It is only run automatically the first time and needs to be
manually called with the `--prepare-redo' option. But this wasn't explained
in `README-hacking.md' (currently the main documentation of Maneage).

With this commit, a description about invoking the preparation process
after the first attempt of the running project has been added to
`README-hacking.md'.