GNU Gettext built as a dependency of Bash
Until now Maneage used the host's GNU Gettext if it was present. Gettext is
a relatively low-level software that enables programs to print messages in
different languages based on the host environment. Even though it has not
direct effect on the running of the software for Maneage and the lanugage
environment in Maneage is pre-determined, it is necessary to have it
because if the basic programs see it in the host they will link with it and
will have problems if/when the host's Gettext is updated.

With this commit (which is actually a squashed rebase of 9 commits by Raul
and Mohammad), Gettext and its two extra dependencies (libxml2 and
libunistring) are now installed within Maneage as a basic software and
built before GNU Bash. As a result, all programs built afterwards will
successfully link with our own internal version of Gettext and
libraries. To get this working, some of the basic software dependencies had
to updated and re-ordered and it has been tested in both GNU/Linux and
macoS.

Some other minor issues that are fixed with this commit

 - Until this commit, when TeX was not installed, the warning message
   saying how to run the configure step in order to re-configure the
   project was not showing the option `-e'. However, the use of this option
   is more convenient than entering the top-build directory and etc every
   time. So with this commit, the warning message has been changed in order
   use the option `-e' in the re-configure of the project.

 - Until now, on macOS systems, Bash was not linking with our internally
   built `libncurses'. With this commit, this has been fixed by setting
   `--withcurses=yes' for Bash's configure script.