Corrected latex citation of sip_tpv program (no more \ before _)
Until now there we had manually inserted a `\' before the `_' of sip_tpv
program. However, we also recently added a step in the configure script to
add a `\' before every `_' when writing the final LaTeX macro. This was
because some C compilers (when the host's is used) have an `_' in their
version that we had no control over.

With this commit, the `\' is removed from `sip_tpv' in its build-rule and
we let the backslash be inserted automatically.