Core software build before using Make to build other software
Until now, Maneage would only build Flock before building everything else
using Make (calling 'basic.mk') in parallel. Flock was necessary to avoid
parallel downloads during the building of software (which could cause
network problems). But after recently trying Maneage on FreeBSD (which is
not yet complete, see bug #58465), we noticed that the BSD implemenation of
Make couldn't parse 'basic.mk' (in particular, complaining with the 'ifeq'
parts) and its shell also had some peculiarities.

It was thus decided to also install our own minimalist shell, Make and
compressor program before calling 'basic.mk'. In this way, 'basic.mk' can
now assume the same GNU Make features that high-level.mk and python.mk
assume. The pre-make building of software is now organized in
'reproduce/software/shell/pre-make-build.sh'.

Another nice feature of this commit is for macOS users: until now the
default macOS Make had problems for parallel building of software, so
'basic.mk' was built in one thread. But now that we can build the core
tools with GNU Make on macOS too, it uses all threads. Furthermore, since
we now run 'basic.mk' with GNU Make, we can use '.ONESHELL' and don't have
to finish every line of a long rule with a backslash to keep variables and
such.

Generally, the pre-make software are now organized like this: first we
build Lzip before anything else: it is downloaded as a simple '.tar' file
that is not compressed (only ~400kb). Once Lzip is built, the pre-make
phase continues with building GNU Make, Dash (a minimalist shell) and
Flock. All of their tarballs are in '.tar.lz'. Maneage then enters
'basic.mk' and the first program it builds is GNU Gzip (itself packaged as
'.tar.lz'). Once Gzip is built, we build all the other compression software
(all downloaded as '.tar.gz'). Afterwards, any compression standard for
other software is fine because we have it.

In the process, a bug related to using backup servers was found in
'reproduce/analysis/bash/download-multi-try' for calling outside of
'basic.mk' and removed Bash-specific features. As a result of that bug-fix,
because we now have multiple servers for software tarballs, the backup
servers now have their own configuration file in
'reproduce/software/config/servers-backup.conf'. This makes it much easier
to maintain the backup server list across the multiple places that we need
it.

Some other minor fixes:

 - In building Bzip2, we need to specify 'CC' so it doesn't use 'gcc'.

 - In building Zip, the 'generic_gcc' Make option caused a crash on FreeBSD
   (which doesn't have GCC).

 - We are now using 'uname -s' to specify if we are on a Linux kernel or
   not, if not, we are still using the old 'on_mac_os' variable.

 - While I was trying to build on FreeBSD, I noticed some further
   corrections that could help. For example the 'makelink' Make-function
   now takes a third argument which can be a different name compared to the
   actual program (used for examle to make a link to '/usr/bin/cc' from
   'gcc'.

 - Until now we didn't know if the host's Make implementation supports
   placing a '@' at the start of the recipe (to avoid printing the actual
   commands to standard output). Especially in the tarball download phase,
   there are many lines that are printed for each download which was really
   annoying. We already used '@' in 'high-level.mk' and 'python.mk' before,
   but now that we also know that 'basic.mk' is called with our custom GNU
   Make, we can use it at the start for a cleaner stdout.

 - Until now, WCSLIB assumed a Fortran compiler, but when the user is on a
   system where we can't install GCC (or has activated the '--host-cc'
   option), it may not be present and the project shouldn't break because
   of this. So with this commit, when a Fortran compiler isn't present,
   WCSLIB will be built with the '--disable-fortran' configuration option.

This commit (task #15667) was completed with help/checks by Raul
Infante-Sainz and Boud Roukema.