IMPORTANT: bug fix in default data download script of download.mk
Summary of possible semantic conflicts
 1. The recipe to download input datasets has been modified. You have to
    re-set the old 'origname' variable to 'localname' (to avoid confusion)
    and the default dataset URL should now be complete (including the
    actual filename). See the newly added descriptions in 'INPUTS.conf' for
    more on this.

Until now, when the dataset was already present on the host system, a link
couldn't be made to it, causing the project to crash in the checksum
phase. This has been fixed with properly naming the main variable as
'localname' to avoid the confusion that caused it.

Some other problems have been fixed in this recipe in the meantime:
 - When the checksum is different, the expected and calculated checksums
   are printed.
 - In the default paper, we now print the full URL of the dataset, not just
   the server, so the checksum of the 'download.tex' step has been updated.