Updated text of default paper.tex, putting more recent examples
The text of the default paper hadn't been changed for a very long time! In
this time, three papers using Maneage have been published (which can be
very good as an example), Maneage also now has a webpage!

With these commit these examples and the webpage have been added and
generally it was also polished a little to hopefully be more useful.