Better comments for the top macros of paper.tex
The default 'paper.tex' starts by defining some macros and comments
describing them. Until now, the text was not too clear and could be
confusing for someone that is not at all familiar with Maneage.

With this commit, the comments have been edited to be more clear for a
first-time reader. For example they all start with FULL CAPS
summaries.

Two other small things were corrected in 'tex/src/preamble-necessary.tex':

 - Until now 'project.tex' was included in this preamble. However, because
   of its importance in Maneage, and prominent place in the demonstration
   plot of the paper introducing Maneage, it is now included directly in
   'paper.tex'. This also allows users to safely ignore/delete this
   preamble file if their LaTeX style is different.

 - I noticed that some macros for some astronomical software names from the
   very first commits in Maneage were still present here! They are no
   longer used, so they have been removed.